pcc-prototype/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
kafka-python==2.0.2
duckdb==0.10.0
faker==22.2.0
python-dateutil==2.8.2
hl7==0.3.5
pytest==7.4.4
httpx==0.26.0
numpy==1.26.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.35.0
python-dotenv==1.0.0